requirements.txt
flask>=2.0
solana>=0.22.0
base58>=2.1.0
pynacl>=1.5.0
pycuda>=2021.1
numpy>=1.20
